separated analysis_lib into multiple files by concerns and to better reflect high level architecture
